requirements.txt
requests
pymupdf
pandas
openpyxl
instructor
openai
pydantic
python-dotenv
